doc: mention transfer data leaks in more places
The "SECURITY" section of the gitnamespaces(7) man page described two
ways for a client to steal data from a server that wasn't intended to be
shared. Similar attacks can be performed by a server on a client, so
adapt the section to cover both directions and add it to the
git-fetch(1), git-pull(1), and git-push(1) man pages. Also add
references to this section from the documentation of server
configuration options that attempt to control data leakage but may not
be fully effective.

Signed-off-by: Matt McCutchen <matt@mattmccutchen.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>
(cherry picked from commit 235ec24352e151bed37063a004b9800ee0debd74)
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>

Gbp-Pq: Name doc-mention-transfer-data-leaks-in-more-places.diff