cherry-pick: demonstrate a segmentation fault
In https://github.com/git-for-windows/git/issues/952, a complicated
scenario was described that leads to a segmentation fault in
cherry-pick.

It boils down to a certain code path involving a renamed file that is
dirty, for which `refresh_cache_entry()` returns `NULL`, and that
`NULL` not being handled properly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>
(cherry picked from commit 05f2dfb965476a59050b7c3446b1281bdcac7051)
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>

Gbp-Pq: Name cherry-pick-demonstrate-a-segmentation-fault.diff