git_exec_path: do not return the result of getenv()
The result of getenv() is not guaranteed by POSIX to last
beyond another call to getenv(), or setenv(), etc.  We
should duplicate the string before returning to the caller
to avoid any surprises.

We already keep a cached pointer to avoid repeatedly leaking
the result of system_path(). We can use the same pointer
here to avoid allocating and leaking for each call.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>
(cherry picked from commit 007ac544011213045e3905983b4350ffec8f41f7)
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>

Gbp-Pq: Name git_exec_path-do-not-return-the-result-of-getenv.diff