apply: use strcmp(3) for comparing strings in gitdiff_verify_name()
We don't know the length of the C string "another".  It could be
shorter than "name", which we compare it to using memchr(3).  Call
strcmp(3) instead to avoid running over the end of the former, and
get rid of a strlen(3) call as a bonus.

Signed-off-by: Rene Scharfe <l.s.r@web.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>
(cherry picked from commit 2d105451c0768fc3e9600dec7bca2376f482521e)
Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>

Gbp-Pq: Name apply-use-strcmp-3-for-comparing-strings-in-gitdiff_v.diff